Tweak a few things about the threadsafe let solution from #1858.
1). Remove `--threadsafe` CLI option.

In general, we don't expose every config option via the CLI.
We want to make it easy for users to run `rspec --help` and
find the options that they are commonly going to want to customize
for a particular CLI run. I don't think `--threadsafe` is one of
those -- it's more something that'll be turned off globally for the
project or not touched at all, and as such, it adds noise to
the `--help` output to include it.

2). Extract Mutex class into its own file.

- Remove need for use of `MUTEX` over `Mutex`.
- No reason to force Ruby to parse the code
  for 1.8.7 Mutex implementation on other Rubies.

3). Remove threadsafe scenario.

A note about threadsafey is sufficient for the docs. Having
a full threadsafety example spec in the scenario is more
detail than users are likely to want to read.
