+++ lib/rspec/core/reentrant_mutex.rb
@@ -14,7 +14,7 @@ module RSpec
       def initialize
         @owner = nil
         @count = 0
-        @mutex = MUTEX.new
+        @mutex = Mutex.new
       end
 
       def synchronize
@@ -40,68 +40,13 @@ module RSpec
       end
     end
 
-    # @private
-    # :nocov:
-    # This can be deleted once support for 1.8.7 is dropped
-    MUTEX = if defined? ::Mutex
-              # On 1.9 and up, this is in core, so we just use the real one
-              ::Mutex
-            else
-              # On 1.8.7, it's in the stdlib.
-              # We don't want to load the stdlib, b/c this is a test tool, and can affect the test environment,
-              # causing tests to pass where they should fail.
-              #
-              # So we're transcribing/modifying it from https://github.com/ruby/ruby/blob/v1_8_7_374/lib/thread.rb#L56
-              # Some methods we don't need are deleted.
-              # Anything I don't understand (there's quite a bit, actually) is left in.
-              # Some formating changes are made to appease the robot overlord:
-              #   https://travis-ci.org/rspec/rspec-core/jobs/54410874
-              Class.new do
-                def initialize
-                  @waiting = []
-                  @locked = false
-                  @waiting.taint
-                  taint
-                end
-
-                def lock
-                  while Thread.critical = true && @locked
-                    @waiting.push Thread.current
-                    Thread.stop
-                  end
-                  @locked = true
-                  Thread.critical = false
-                  self
-                end
-
-                def unlock
-                  return unless @locked
-                  Thread.critical = true
-                  @locked = false
-                  begin
-                    t = @waiting.shift
-                    t.wakeup if t
-                  rescue ThreadError
-                    retry
-                  end
-                  Thread.critical = false
-                  begin
-                    t.run if t
-                  rescue ThreadError
-                    :noop
-                  end
-                  self
-                end
-
-                def synchronize
-                  lock
-                  begin
-                    yield
-                  ensure
-                    unlock
-                  end
-                end
-              end
-            end
+    if defined? ::Mutex
+      # On 1.9 and up, this is in core, so we just use the real one
+      Mutex = ::Mutex
+    else # For 1.8.7
+      # :nocov:
+      RSpec::Support.require_rspec_core "mutex"
+      # :nocov:
+    end
   end
 end
